Start GetFlows before generating traffic
Start GetFlows with the follow option in a separate goroutine before
generating traffic to reduce the chance we miss flows.

This PR only modifies connectivity/tests/world.go, which seems to be
causing the most flow validation failures. We can update other tests
to follow the same pattern once we confirm this improves the overall
stability of flow validation.

Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>